(MainHeader、loginPopupChange)登入驗證+登出、取得會員資料api、樣式微調

+++ src/router/index.js
@@ -60,25 +60,7 @@ const routes = [
         props: true
       }
     ]
-    // children:[
-    //   {
-    //     path:'',
-    //     component: () => import('@/pages/Social_news.vue'),
-    //   },
-    //   {
-    //     path: 'article', // 'article/:id'
-    //     component:  () => import('@/pages/Social_article.vue'),
-    //   },
-    // ]
   },
-  // {
-  //   path: '/social_article/',
-  //   component: () => import('@/pages/Social_article.vue'),
-  //   meta: {
-  //     title: '社群中心_文章',
-  //     // requiredLogin: true
-  //   },
-  // },
   {
     path: '/About/',
     component: () => import('@/pages/About.vue'),
@@ -109,13 +91,6 @@ const routes = [
       requiredLogin: true
     },
   },
-  // {
-  //   path: '/social_news',
-  //   component: () => import('@/pages/Social_news.vue'),
-  //   meta: {
-  //     title: '新聞',
-  //   },
-  // },
   {
     path: '/popup/',
     component: () => import('@/pages/popup.vue'),
@@ -310,56 +285,20 @@ router.beforeEach(async (to, from, next) => {
 
     try {
       const isAuthenticated = await checkAuth();
-
       if (!isAuthenticated) {
-        localStorage.setItem('redirectPath', to.fullPath); //存儲原本要去的頁面
-        eventBus.emit('show-login-popup'); //跳出登入彈窗
+        const askLogin = confirm('此頁面需要先登入再進行操作，請問是否繼續？');
+        if(askLogin){
+          localStorage.setItem('redirectPath', to.fullPath); //存儲原本要去的頁面
+          eventBus.emit('show-login-popup'); //跳出登入彈窗
+        }
         next(false)
       } else {
         next()
       }
-
-
-
-      // const res = await fetch('/tid103/g1/php/checkLogin.php',{
-      //   credentials: 'include'
-      // });
-      // const data = await res.json();
-      // if(!data.success){
-      //   // next({
-      //   //   path:'/',
-      //   //   query: { redirect: to.fullPath } // 保存原本要去的路徑
-      //   // });
-
-      //   //顯示登入的燈箱...
-
-      //   next(false);
-      // }else{
-      //   next();
-      // }
-
     } catch (err) {
       console.error('檢查登入狀態失敗:', err);
       next(false);
     }
-    // ======= 以下要取得使用者目前的登入狀態，會是 bool == //
-    // 取得是否已登入，可能是從 localStorage 抓資料或從後端判斷。
-    // let isAuthenticated = true;
-    // const res = await fetch('https://notes.webmix.cc/login_test/login.php');
-    // const data = await res.json();
-
-    // console.log(data);
-
-    // const isAuthenticated = data.login_status;
-    // // ============================================== //
-
-    // if (isAuthenticated) {
-    //   document.title = to.meta.title;
-    //   next();
-    // } else {
-    //   // 未登入，就直接導回到首頁或其它頁面。
-    //   next('/');
-    // }
   } else {
     document.title = to.meta.title;
     next();
